Ensure eval test functions fork contextes and rename exceptional
Forking a context prior to testing a condition is safer. Furthermore,
there is now a notion of a base context used for tests which can be
prepared as needed (eg. add utilities).

Function related to exceptional states now refer to errors, avoiding
confusion between CVM exceptionals and actual JVM exceptions.