Prototype strxs as regular vectors
Instead of checking proactively is a transaction is a special one,
before execution, the result after execution is checked. It must be a
tuple vector following a particular format given each special
transaction.

As such, they are more flexible and macros can be written which actually
outputs those specials vectors and do argument checking.

For the time being, transaction hooks and dependency manaagement are
disabled.