Reimplement trx hooks in runner
Now, a transaction hook is a transaction meant to be executed after each
regular transaction.

Before, it was a function `trx code` -> `trx code` but it has proven to
be difficult to handle and counterproductive. The main purpose of a
transaction hook is to provide feedback on what happened (eg. report
consumed juice).