Remove `:project/watch`
The CVX file watcher was primarily implemented for the CVX runner.
However, in a recent commit, it has been decided that file watching was
not the right approach given simple file loading and a decent REPL
experience.

File watching becomes a burden for more complex workflow. Hence, since
it is not recommend, the related codebase is simply removed.