Move `:project/cvx` to core `convex` repository
The purpose behind developping and testing officials Convex actors
elsewhere than the core repository was to leverage Clojure tooling.
However, the runner became so convenient that this idea became obsolete,
especially since it will support generative testing in pure CVX in a
future iteration.

It makes more sense to keep those files colocated with the CVM.