Make CVX Shell requests stand out
Previously, requests were named as regular functions. However, this was
probably confusing for beginners since results are delayed to the next
transactions.

In order to make that fact more explicit, request names are now
prepended with `!.`. E.g. `$.stream/!.outln`. At the moment, this is
true of all shell libraries but `$.test`.