Remove CVX Shell utilities for appending transactions
Appending to a list (of transactions) is not very efficient (in theory).
Furthermore, after all this complex CVX code written for the CVX Shell,
it has been proven empirically it is never needed while prepending is
very useful.

Users can still use `$.trx/set` to do any kind of arbitrary operation on
the list of transactions.